Made a start on turning this into a proper CLI app
Overall architecture of the CLI version will have to be based around a
daemon for running the repository server, background filesystem
monitoring and the downloading client. A GUI client can easily integrate
all of that functionality into a single process, as we can generally
expect the GUI to keep running.

And with this architecture it's easier to keep all user interaction
through the CLI, so I don't have to spend much effort on a convenient
human-editable config file format. JSON will do fine.

This commit really just lays out the basic framework, which will
definitely be rewritten a few more times and doesn't actually *do*
anything yet. So lots of effort for basically nothing.